Adds a new View and its related ViewModel
We've added a new view to serve us as a Sales Page, responsible for the checkout process inside a store where the clerk will check the purchased products and pass the bill for the costumer to him to make the payment.

We've not implement any code yet just the XAML markupt and wire it up (the view) with it's related viewModel.